13. gün ödev

+++ Business/Concrete/ImageManager.cs
@@ -2,6 +2,8 @@
 using Business.Constants;
 using Business.ValidationRules.FluentValidation;
 using Core.Aspects.Autofac.Validation;
+using Core.Utilities.Business;
+using Core.Utilities.Helpers;
 using Core.Utilities.Results;
 using DataAccess.Abstract;
 using Entities.Concrete;
@@ -9,27 +11,26 @@ using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Text;  
-using Core.Utilities.Business;
-using Core.Utilities.Helpers;
 
 namespace Business.Concrete
 {
     public class ImageManager : IImageService
     {
         IImageDal _imageDal;
-        public ImageManager(IImageDal imageDal)
+        ICarDal _carDal;
+        public ImageManager(IImageDal imageDal, ICarDal carDal)
         {
             _imageDal = imageDal;
+            _carDal = carDal;
         }
 
-        //[ValidationAspect(typeof(ImageValidator))]
+        [ValidationAspect(typeof(ImageValidator))]
         public IResult Add(IFormFile file, Image image)
         {
-            var result = BusinessRules.Run(CheckImageLimit(image.CarId));
-            if (result != null)
+            var check = BusinessRules.Run(CheckImageLimit(image.CarId),CheckTheCarExists(image.CarId));
+            if (check != null)
             {
-                return result;
+                return check;
             }
 
             image.ImagePath = FormFileHelper.Add(file);
@@ -40,7 +41,11 @@ namespace Business.Concrete
 
         public IResult Delete(Image image)
         {
-
+            var check = BusinessRules.Run(CheckTheCarExists(image.CarId));
+            if (check != null)
+            {
+                return check;
+            }
 
             FormFileHelper.Delete(image.ImagePath);
             _imageDal.Delete(image);
@@ -54,11 +59,22 @@ namespace Business.Concrete
 
         public IDataResult<List<Image>> GetImagesByCarId(int carId)
         {
+            var check = BusinessRules.Run(CheckTheCarExists(carId));
+            if (check != null)
+            {
+                return new ErrorDataResult<List<Image>>(Messages.CarDoesntExists);
+            }
             return new SuccessDataResult<List<Image>>(_imageDal.GetAll(i => i.CarId == carId));
         }
 
         public IResult Update(IFormFile file, Image image)
         {
+            var check = BusinessRules.Run(CheckTheCarExists(image.CarId));
+            if (check != null)
+            {
+                return check;
+            }
+
             var result = Path.GetFullPath(Path.Combine(AppContext.BaseDirectory, "../../../wwwroot")) +
                          _imageDal.Get(i => i.ImageId == image.ImageId).ImagePath;
             image.ImagePath = FormFileHelper.Update(file, result);
@@ -73,6 +89,16 @@ namespace Business.Concrete
             if (result.Count > 5)
             {
                 return new ErrorResult(Messages.ImageAdditionFailed);
+            }
+            return new SuccessResult();
+        }
+
+        private IResult CheckTheCarExists(int carId)
+        {
+            var check = _carDal.GetAll(c => c.Id == carId);
+            if (check.Count == 0)
+            {
+                return new ErrorResult(Messages.CarDoesntExists);
             }
             return new SuccessResult();
         }
